+++ app/src/pages/Projects/allProjectsGQL.js
@@ -4,6 +4,7 @@ const generalData = gql`
   fragment generalData on Project {
     id
     name
+    slug
     description
     ticker
     coinmarketcapId
@@ -29,6 +30,7 @@ const project = gql`
     ethSpent
     averageDevActivity
     averageDailyActiveAddresses
+    fundsRaisedUsdIcoEndPrice
     marketcapUsd
     ethBalance
     btcBalance
@@ -47,6 +49,17 @@ const project = gql`
 export const allProjectsGQL = gql`
   query allProjects {
     allProjects {
+      ...generalData
+      ...project
+    }
+  }
+  ${generalData}
+  ${project}
+`
+
+export const projectBySlugGQL = gql`
+  query projectBySlugGQL($slug: String!) {
+    projectBySlug(slug: $slug) {
       ...generalData
       ...project
     }
@@ -129,7 +142,28 @@ export const allShortProjectsGQL = gql`
 
 export const allMarketSegmentsGQL = gql`
   {
-    allMarketSegments
+    allMarketSegments {
+      name
+      count
+    }
+  }
+`
+
+export const erc20MarketSegmentsGQL = gql`
+  {
+    erc20MarketSegments {
+      name
+      count
+    }
+  }
+`
+
+export const currenciesMarketSegmentsGQL = gql`
+  {
+    currenciesMarketSegments {
+      name
+      count
+    }
   }
 `