+++ app/src/reducers/detailed-page-ui.js
@@ -38,10 +38,6 @@ export default (state = initialState, action) => {
           interval
         }
       }
-    case 'TOGGLE_FOLLOW':
-      return {
-        ...state
-      }
     default:
       return state
   }