+++ app/src/utils/utils.spec.js
@@ -80,32 +80,17 @@ describe('sanitizeMediumDraftHtml', () => {
 
 describe('filterProjectsByMarketSegment', () => {
   it('should return expected values', () => {
-    const allMarketSegments = {
-      advertising: 'Advertising',
-      blockchain_network: 'Blockchain Network',
-      data: 'Data',
-      digital_identity: 'Digital Identity',
-      financial: 'Financial',
-      gambling: 'Gambling',
-      gaming: 'Gaming',
-      legal: 'Legal',
-      media: 'Media',
-      prediction_market: 'Prediction Market',
-      protocol: 'Protocol',
-      transportation: 'Transportation',
-      unknown: null
-    }
     const dataProvider = [
       // projects is undefined, there is a category
       {
         projects: undefined,
-        categories: { financial: true },
+        categories: { Financial: true },
         expectation: undefined
       },
       // projects is an empty array, there is a category
       {
         projects: [],
-        categories: { financial: true },
+        categories: { Financial: true },
         expectation: []
       },
       // projects is an empty array, categories is an empty object
@@ -123,7 +108,7 @@ describe('filterProjectsByMarketSegment', () => {
       // projects is not empty, there is a category
       {
         projects: [{ marketSegment: 'Financial' }, { marketSegment: 'media' }],
-        categories: { financial: true },
+        categories: { Financial: true },
         expectation: [{ marketSegment: 'Financial' }]
       },
       // projects is not empty, there are multiple categories
@@ -134,8 +119,8 @@ describe('filterProjectsByMarketSegment', () => {
           { marketSegment: 'Advertising' }
         ],
         categories: {
-          financial: true,
-          blockchain_network: true
+          Financial: true,
+          'Blockchain Network': true
         },
         expectation: [
           { marketSegment: 'Financial' },
@@ -145,11 +130,7 @@ describe('filterProjectsByMarketSegment', () => {
     ]
     dataProvider.map(data =>
       expect(
-        filterProjectsByMarketSegment(
-          data.projects,
-          data.categories,
-          allMarketSegments
-        )
+        filterProjectsByMarketSegment(data.projects, data.categories)
       ).toEqual(data.expectation)
     )
   })