+++ app/stories/SearchStory.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import Search from './../src/components/Search'
+import Search from './../src/components/Search/Search'
 
 const projects = [{
   name: 'Santiment',