+++ lib/sanbase_web/graphql/resolvers/tech_indicators_resolver.ex
@@ -98,6 +98,68 @@ defmodule SanbaseWeb.Graphql.Resolvers.TechIndicatorsResolver do
     )
   end
 
+  def erc20_exchange_funds_flow(
+        _root,
+        %{
+          from: from,
+          to: to
+        },
+        _resolution
+      ) do
+    TechIndicators.erc20_exchange_funds_flow(
+      from,
+      to
+    )
+  end
+
+  def social_volume(
+        _root,
+        %{
+          slug: slug,
+          from: from,
+          to: to,
+          interval: interval,
+          social_volume_type: social_volume_type
+        },
+        _resolution
+      ) do
+    TechIndicators.social_volume(
+      slug,
+      from,
+      to,
+      interval,
+      social_volume_type
+    )
+  end
+
+  def social_volume_projects(
+        _root,
+        %{},
+        _resolution
+      ) do
+    TechIndicators.social_volume_projects()
+  end
+
+  def topic_search(
+        _root,
+        %{
+          sources: sources,
+          search_text: search_text,
+          from: from,
+          to: to,
+          interval: interval
+        },
+        _resolution
+      ) do
+    TechIndicators.topic_search(
+      sources,
+      search_text,
+      from,
+      to,
+      interval
+    )
+  end
+
   defp price_volume_diff_ma_window_type() do
     Config.module_get(Sanbase.Notifications.PriceVolumeDiff, :window_type)
   end